Added controllers test
